+++ maseval/__init__.py
@@ -16,6 +16,7 @@ from .core.task import (
     BaseTaskQueue,
     TaskQueue,
     SequentialTaskQueue,
+    AnchorPointsTaskQueue,
     PriorityTaskQueue,
     AdaptiveTaskQueue,
 )
@@ -93,6 +94,7 @@ __all__ = [
     "BaseTaskQueue",
     "TaskQueue",
     "SequentialTaskQueue",
+    "AnchorPointsTaskQueue",
     "PriorityTaskQueue",
     "AdaptiveTaskQueue",
     # Model adapters